refactor(config): consolidate Keycloak config into keycloakConfig.js
- Added KEYCLOAK_CONFIG and DEMO_CONFIG as named exports in keycloakConfig.js
- Updated useFeatureFlags.js to import from keycloakConfig.js
- Deleted config/environment.js (duplicate of keycloakConfig.js base values)

Previously both files defined the same Keycloak URL, realm, and clientId.
Now there is a single source of truth for all Keycloak-related configuration.
